requirements.txt
openai==1.12.0
praw==7.7.1
langchain==0.1.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.5.2
tenacity==8.2.3
langchain-community==0.0.13
langchain-core==0.1.12
langchain-openai==0.0.5